backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Flask-CORS==4.0.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
email-validator==1.1.3
Werkzeug==2.0.1
Flask-Mail==0.9.1
fhirclient==3.0.0
requests==2.26.0
python-jose==3.3.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
PyJWT==2.8.0
pandas==2.0.3
bcrypt==4.0.1 
twilio==8.10.0
gunicorn==20.1.0 